modify the testing strategy
