+++ zebra-state/src/service/finalized_state/tests.rs
@@ -1,1 +1,1 @@
-mod vectors;
+mod prop;
